load the config through an external file
There are modules to handle config settings in a more agnostic way.  I'm personally familiar with env style configs, and like them, but for the sake of keeping it simple, decided to load through a js file.  There has been a lot of dialogue about which method is best, but ultimately the reasoning and layout for js objects won out for me.