added value checking for the config
this needs more testing, and the regexs could stand to be fine tuned more.
also, the general loops should probably be re-thought some more.
but this is working.

changed the hostname key to interface, since you can bind by IP or hostname.
in this case, binding by IP seems more appropriate, even if node is doing a lookup on hostname.

also added a requirement for the api_key value, to be 16-40 length alphanumeric characters.
perhaps there should be a configurable setting to not use the key, but will think that through later.

still need to update the documentation for the changes.